// EscalafonFrontend/src/app/interfaces/capacitacion.ts
export interface Capacitacion {
    /**
     * @type { number } - iCapaId
     */
    iCapaId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iEduTipEstId
     */
    iEduTipEstId?: number;

    /**
     * @type { Date } - dtFechaInicio
     */
    dtFechaInicio?: Date;

    /**
     * @type { Date } - dtFechaFin
     */
    dtFechaFin?: Date;

    /**
     * @type { number } - iEduTipPartId
     */
    iEduTipPartId?: number;

    /**
     * @type { string } - cCapaInstitucion
     */
    cCapaInstitucion?: string;

    /**
     * @type { number } - iPaisId
     */
    iPaisId?: number;

    /**
     * @type { string } - cCapaCiudad
     */
    cCapaCiudad?: string;

    /**
     * @type { string } - cCapaTema
     */
    cCapaTema?: string;

    /**
     * @type { number } - iEduDocAcredId
     */
    iEduDocAcredId?: number;

    /**
     * @type { string } - cCapaNumeroRegistro
     */
    cCapaNumeroRegistro?: string;

    /**
     * @type { Date } - dtCapaFechaEmision
     */
    dtCapaFechaEmision?: Date;

    /**
     * @type { number } - iCapaDuracionHoras
     */
    iCapaDuracionHoras?: number;

    /**
     * @type { number } - nCapaCreditos
     */
    nCapaCreditos?: number;

    /**
     * @type { number } - iEduSemId
     */
    iEduSemId?: number;

    /**
     * @type { number } - iEduModaId
     */
    iEduModaId?: number;

    /**
     * @type { string } - cCapaAnotaciones
     */
    cCapaAnotaciones?: string;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

}