// EscalafonFrontend/src/app/interfaces/v-capacitacion.ts
export interface VCapacitacion {
    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iEduTipEstId
     */
    iEduTipEstId?: number;

    /**
     * @type { Date } - dtFechaInicio
     */
    dtFechaInicio?: Date;

    /**
     * @type { Date } - dtFechaFin
     */
    dtFechaFin?: Date;

    /**
     * @type { number } - iEduTipPartId
     */
    iEduTipPartId?: number;

    /**
     * @type { number } - iEduDocAcredId
     */
    iEduDocAcredId?: number;

    /**
     * @type { number } - iPaisId
     */
    iPaisId?: number;

    /**
     * @type { string } - cCapaCiudad
     */
    cCapaCiudad?: string;

    /**
     * @type { string } - cCapaNumeroRegistro
     */
    cCapaNumeroRegistro?: string;

    /**
     * @type { Date } - dtCapaFechaEmision
     */
    dtCapaFechaEmision?: Date;

    /**
     * @type { number } - iCapaDuracionHoras
     */
    iCapaDuracionHoras?: number;

    /**
     * @type { number } - iEduSemId
     */
    iEduSemId?: number;

    /**
     * @type { number } - nCapaCreditos
     */
    nCapaCreditos?: number;

    /**
     * @type { number } - iEduModaId
     */
    iEduModaId?: number;

    /**
     * @type { string } - cCapaAnotaciones
     */
    cCapaAnotaciones?: string;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { string } - cEduTipEstNombre
     */
    cEduTipEstNombre?: string;

    /**
     * @type { string } - cCapaTema
     */
    cCapaTema?: string;

    /**
     * @type { string } - cEduDocAcredNombre
     */
    cEduDocAcredNombre?: string;

    /**
     * @type { string } - cCapaInstitucion
     */
    cCapaInstitucion?: string;

    /**
     * @type { number } - iCapaId
     */
    iCapaId?: number;

}